Refactor hit test handling
