+++ global.json
@@ -1,6 +1,6 @@
 {
   "sdk": {
-    "version": "8.0.117",
+    "version": "9.0.100",
     "rollForward": "latestMinor",
     "allowPrerelease": true
   }
