+++ samples/TestApp/Views/MainView.axaml.cs
@@ -5,11 +5,9 @@ using System.Diagnostics;
 using System.Runtime.InteropServices;
 using Avalonia.Input;
 using Avalonia.Interactivity;
-using Avalonia.VisualTree;
 using ShimSkiaSharp;
 using System.Collections.ObjectModel;
 using Avalonia;
-using Avalonia.Media;
 using TestApp.ViewModels;
 
 namespace TestApp.Views;
@@ -24,11 +22,6 @@ public partial class MainView : UserControl
         AddHandler(DragDrop.DropEvent, Drop);
         AddHandler(DragDrop.DragOverEvent, DragOver);
         HitResults.ItemsSource = _hitResults;
-    }
-
-    private void InitializeComponent()
-    {
-        AvaloniaXamlLoader.Load(this);
     }
 
     private void DragOver(object? sender, DragEventArgs e)
@@ -97,37 +90,20 @@ public partial class MainView : UserControl
 
     private void Svg_OnPointerPressed(object? sender, PointerPressedEventArgs e)
     {
-        var svg = Svg.SkSvg;
-        if (svg?.Picture is null)
-        {
-            return;
-        }
-
         var pt = e.GetPosition(Svg);
 
-        var picture = svg.Picture;
-        var viewPort = new Rect(Svg.Bounds.Size);
-        var sourceSize = new Size(picture.CullRect.Width, picture.CullRect.Height);
-        var scale = Svg.Stretch.CalculateScaling(Svg.Bounds.Size, sourceSize, Svg.StretchDirection);
-        var scaledSize = sourceSize * scale;
-        var destRect = viewPort.CenterRect(new Rect(scaledSize)).Intersect(viewPort);
-        var sourceRect = new Rect(sourceSize).CenterRect(new Rect(destRect.Size / scale));
-        var bounds = picture.CullRect;
-        var scaleMatrix = Matrix.CreateScale(destRect.Width / sourceRect.Width, destRect.Height / sourceRect.Height);
-        var translateMatrix = Matrix.CreateTranslation(-sourceRect.X + destRect.X - bounds.Top, -sourceRect.Y + destRect.Y - bounds.Left);
-        var matrix = scaleMatrix * translateMatrix;
-        var inverse = matrix.Invert();
-        var picturePoint = inverse.Transform(pt);
-
-        var skPoint = new SKPoint((float)picturePoint.X, (float)picturePoint.Y);
-
         _hitResults.Clear();
-        svg.Settings.HitTestPoints.Clear();
-        svg.Settings.HitTestPoints.Add(skPoint);
-        foreach (var element in svg.HitTestElements(skPoint))
+
+        if (Svg.SkSvg is { } skSvg && Svg.TryGetPicturePoint(pt, out var skPoint))
         {
-            var id = element.ID ?? element.GetType().Name;
-            _hitResults.Add(id);
+            skSvg.Settings.HitTestPoints.Clear();
+            skSvg.Settings.HitTestPoints.Add(skPoint);
+
+            foreach (var element in Svg.HitTestElements(pt))
+            {
+                var id = element.ID ?? element.GetType().Name;
+                _hitResults.Add(id);
+            }
         }
 
         Svg.InvalidateVisual();
